Rename Button interface to FAButton interface
