+++ src/Components/Atoms/FAButton/index.tsx
@@ -1,7 +1,7 @@
 import React, { memo } from "react";
 import { Button } from "@mui/material";
 
-import { IFAButtonProps } from "./Button.interface";
+import { IFAButtonProps } from "./FAButton.interface";
 
 export const FAButton = memo((props: IFAButtonProps) => {
   const { testID, ...materialUIProps } = props;
